Move deriver instances to deriver package object
